fix: some duplicated types
