+++ packages/job-worker/package.json
@@ -55,7 +55,7 @@
 		"superfly-timeline": "^8.3.1",
 		"threadedclass": "^1.1.2",
 		"tslib": "^2.4.0",
-		"type-fest": "^3.1.0",
+		"type-fest": "^2.19.0",
 		"underscore": "^1.13.4",
 		"vm2": "^3.9.11"
 	},
